requirements.txt
Flask==3.0.3
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Bootstrap
stripe
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
psycopg2-binary
